+++ CHANGELOG.md
@@ -4,13 +4,15 @@
 
 - Removed deprecated namespaces/function/macros:
     - Removed `reagent.interop` namespace (macros `$`, `$!`, `unchecked-aget` and `unchecked-aset`)
-- Deprecated functions:
     - `reagent.core/component-path` (Reason: implementation depends on internal React
     details and using just Component `displayName` achieves nearly the same)
 - All DOM related functions (notably `render` and `dom-node`) have been removed
 from `reagent.core` namespace and are now only available in `reagent.dom` namespace.
 This is to make non-DOM environments (React-native) first class targets with Reagent,
 as requiring `react-dom` always causes problems in such environments.
+- `Error rendering component` messages no longer contain component stack information.
+Instead one should use [an Error Boundary](https://reactjs.org/docs/error-boundaries.html#component-stack-traces)
+to catch the problem and look at the error information `componentStack` property.
 
 ## 0.9.1 (2020-01-15)